[Austin] Restructure report form
