+++ client/src/Forms/ReportForm/index.js
@@ -1,0 +1,3 @@
+import ReportFormContainer from './container'
+
+export default ReportFormContainer
